rename: homepage -> home으로 수정
